Allow vim colorscheme to be changed remotely vim/colorscheme/day-colorscheme-name.txt, vim/colorscheme/night-colorscheme-name.txt - These two files contain the word "light" and "dark" respectively. They will be used to tell vim which colorscheme to load.
vim/colorscheme-name.txt - This file is a symlink to either the day or night txt file. This way
we can change which colorscheme without changing the file vim reads from, we just change where the
symlink points.

vim/vimrc - Added a function, SyncColorscheme(), to set the colorscheme based on the contents of a file
(colorscheme-name.txt). Also added an autocmd so that when vim receives the SIGUSR1 signal, it calls
SyncColorscheme().

bin/switch-theme - This is where the colorscheme gets changed. It does this by first changing where
the symlink colorscheme-name.txt points. After that, it send vims the SIGUSR1 signal at which point
vim sets the colorscheme.

install.conf.yaml - Added a symlink for colorscheme-name.txt

.gitignore - Added colorscheme-name.txt here since its just a symlink