(tmux) launch new tmux shells as login shells
Until I read more about login shells, I'm not going to change the default.

Also, when I have a default_command set it screws up tmux-resurrect because it thinks the program I was running in that pane is the one I started with my default_command, but that's just my shell. What I want to restore is the program started from that shell.