(neovim) fix: colorscheme flash on startup
By waiting until VimEnter to set the colorscheme, the default colorscheme has already been loaded so there is a flash when it changes from default to the one I set. There is no need to wait until VimEnter since we know that the colorscheme plugin will be loaded by the time we try to set it.