fix: fish autoreload exits upgrade-all early
When I run the upgrade-all fish function, it would exit whenever I finished the upgrade-fish part. This is because upgrading fish causes a change to the configuration directory of fish which causes fish to autoreload with `exec fish`. This exec call end my upgrade-all function early. I would think that it would wait for the function to finish before autoreloading, but it doesn't. By moving the function to a separate executable, fish now waits for it to finish. I think it's because fish functions are run in the fish shell process, but separate executables are run as a child process.

Since git hooks will be run in a child process of the shell, fish autoreload won't interrupt I removed the warning.