(brew) don't link pkg-config
I have another pkg-config installed in usr/bin, but brew's pkg-config will show up first on the path. Since they are installed in different places, their search paths are different so I can't use the brew one to install things outside of brew. Plus, brew's pkg-config is only needed to install brew packages so there is no need for it to be on the $PATH.