(login-shell) fix: make sure login shell config gets loaded
When logging in with X11, /bin/sh is used to source startup scripts instead of whatever is in $SHELL. One of the startup scripts is ~/.profile so I've moved all the environment configuration into that file. I also added a `bash_profile` that sources `profile`. This way, the environment will still get configured even when bash is used.

source: https://unix.stackexchange.com/a/88149