(home-manager) fix: copy shell output script into the default package
The `program` for the shell (apps.default) output is just a shell script now instead of fish. Unlink the fish binary, shell scripts are not stored in a directory in the nix store, it's just `/nix/<hash>-<script-name>`. `symlinkJoin` only takes directories in the nix store so I just copied the shell script
in during the `postBuild` hook for the package created by `symlinkJoin`.