define inline scripts using `writeShellApplication` and add them as packages.
This avoids boilerplate like specifying where to link it or setting it as an executable. Also scripts defined with `writeShellApplication` get linted by `shellcheck`.
I also moved a `repository.git.onChange` definition into the right spot, we want to define it even if home manager is a submodule because the out system manager won't do it.